pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evmbench-certora-harness"
version = "0.1.0"
description = "Iterative LLM + Certora harness for EVMBench smart contract security tasks"
readme = "README.md"
authors = [{name = "hpagent"}]
requires-python = ">=3.9"
dependencies = [
  "pyyaml>=6.0",
  "requests>=2.31.0",
]

[project.scripts]
evmbench-certora-harness = "evmbench_certora_harness.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
